Changed backend configuration
Not all backends have a single "address" or have a concept of skipping
TLS validation. Therefore, these properties were moved back into the
backend type specific configuration. Also, the backend type specific
configuration was changed to be called "properties". The backend
configuration also technically now takes a property called "name" (even
though it is not yet being used) in preparation for supporting multiple
backends.

This all came with a bit of a refactor, because pulling all of these
values out of a map and type checking them individually was making it
difficult to follow the code. Now the values get unmarshalled into a
struct.

These changes still need to be updated in the documentation at this
point.