Refactor server auth config schemas
When the server auth configurations were made, they were implemented as
map[string]string (because time constraints). This presented issues with
having the code be self-documenting - but more importantly, not all
properties are naturally strings (booleans and numbers are nice). So
now, properties get put into a map[string]interface{}, and then
reevaluated into an auth backend-specific struct. Also, the key "config"
was changed to "properties" to be consistent with that of the ingest
backend configuration.