added show more button
This moves the cert data out into an external data structure from the
handler that populates it. This allows page repainting to be done
decoupled from the timing of the polling updates. Then, implemented a
show more/less button that will show more than just the certs expiring
within 90 days.