Vault token renewal and approle
The Vault backend now supports renewable tokens and AppRoles that
authenticate to generate renewable tokens.

There are still shortcomings with this implementation - for one, it
can't really handle a one-off failure, a shortcoming that it shares with
the UAA-authed backends. It also can't handle an AppRole where the token
generated is not infinitely renewable. However, pending the
authentication code getting integrated into the scheduler for a more
robust solution, this is a nice stopgap that offers a lot more
functionality than only supporting root tokens.