Prioritize ready>not ready and auth>refresh
Auth is typically more time sensitive, so we should prioritize it over
refresh.
Also, there was a potential race condition where the task waking up the
next() thread may not have been the front of the queue, and could've
been behind a task that has not quite yet been marked as ready.