Implement task skip logic
A task will be marked as to-skip if, when it would be marked as ready,
there is already another of the same task which is ready or running. A
task which is marked as to-skip is still picked up by the scheduler, but
it is not run. This is so that it runs through the same schedule-next
path as a task that should run, such that a scheduled task doesn't proc
a future scheduled task (that would be bad).